requirements.txt
aiohttp==3.7.4
async-timeout==3.0.1
attrs==20.3.0
beautifulsoup4==4.9.3
bitlyshortener~=0.5.0
cachetools==4.2.1
certifi==2020.12.5
chardet==3.0.4
discord.py==1.6.0
idna==2.10
multidict==5.1.0
mysql-connector-python~=8.0.22
Pillow~=8.0.1
protobuf==3.15.3
psutil==5.8.0
python-dotenv==0.15.0
requests~=2.25.0
six==1.15.0
soupsieve==2.2
typing-extensions==3.7.4.3
urllib3==1.26.3
yarl==1.6.3
discord~=1.0.1
mysql~=0.0.2
bs4~=0.0.1
discord.py